엔티í유저 로êµ그인 및 회원가입 기능 & Auth 수정 중

+++ src/main/java/com/johyeonjung/blogfeed/controller/CommentController.java
@@ -14,13 +14,13 @@ public class CommentController {
     private final CommentService commentService;
 
     @PostMapping
-    public ResponseEntity<?> createComments(@PathVariable int postId, @RequestBody CreatePostCommentReqDto dto) {
+    public ResponseEntity<?> createComments(@PathVariable Long postId, @RequestBody CreatePostCommentReqDto dto) {
         commentService.createComment(postId, dto);
         return ResponseEntity.ok("댓글 작성 완료");
     }
 
     @GetMapping
-    public ResponseEntity<?> getComments(@PathVariable int postId) {
+    public ResponseEntity<?> getComments(@PathVariable Long postId) {
         return ResponseEntity.ok(commentService.getComments(postId));
     }
 }
